Service - add getHeroesSlowly() method
Let’s take stock of what we’ve built.

We created a service class that can be shared by many components.
We used the ngOnInit Lifecycle Hook to get our heroes when our
AppComponent activates.
We defined our HeroService as a provider for our AppComponent.
We created mock hero data and imported them into our service.
We designed our service to return a Promise and our component to get
our data from the Promise.

Like getHeroes, it also returns a Promise. But this Promise waits 2
seconds before resolving the Promise with mock heroes.